fix(releases): consolidate Author union into release_details_types
`release_details_types.Author` was a hand-rolled duplicate of the `UserSerializerResponse | NonMappableUser` union that `get_users_for_authors()` already returns from `models/release.py`. It declared `id: int` (runtime emits `str(obj.id)`) and made 8 fields required that the serializer omits for unmatched git committers. The published `OrganizationReleaseDetailsEndpoint` schema therefore advertised a shape never produced; clients codegen'd against it break on `authors[].id` typing or on any release with an external git committer.


Move the canonical `Author` and `NonMappableUser` definitions from `models/release.py` into `release_details_types.py` (the type-only module `models/release.py` already imports `VersionInfo` from). `models/release.py`, `models/commit.py`, `models/pullrequest.py`, and `utils/committers.py` import them back from there. Single source of truth, no rename — the union is the same shape regardless of commit/PR/release context.


Importing `UserSerializerResponse` here creates a cycle via `api/serializers/types.py`. Break it by moving `SerializedAvatarFields` out of `types.py` into `users/api/serializers/user.py`, where it's primarily used as `UserSerializerResponse.avatar`. Updates the `team`/`project`/`organization` serializer import paths and bumps the existing `RetrieveReleaseDetails` example (`authors[0].id` int → str) so the now-truthful schema validates.
