+++ src/sentry/apidocs/examples/organization_examples.py
@@ -772,7 +772,7 @@ class OrganizationExamples:
                 },
                 "authors": [
                     {
-                        "id": 2837091,
+                        "id": "2837091",
                         "name": "Raj's Raspberries",
                         "username": "rajraspberry",
                         "email": "raj@raspberries",
